refactor: APIモードに変更． #16
フロントエンドをReactで構築するため．
