+++ app/controllers/api/v1/notices_controller.rb
@@ -1,0 +1,15 @@
+module Api
+  module V1
+    class NoticesController < ApplicationController
+      def index
+        @notices = current_user.passive_notices.preload(:sender, :group).where.not(sender_id: current_user.id).order(id: :DESC)
+      end
+
+      def update
+        notice = Notice.find(params[:id])
+        notice.update(checked: params[:checked])
+        redirect_to notices_path
+      end
+    end
+  end
+end
